requirements.txt
langgraph
pydantic
google-generativeai
requests
python-dotenv
yt-dlp
youtube-transcript-api
moviepy
pypdf 
langchain-community
langchain-google-genai
chromadb
fastapi
uvicorn
langchain-chroma
